requirements.txt
watchdog>=3.0.0
PyYAML>=6.0
